core/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.24.0
pydantic>=2.9.0
python-dotenv>=1.0.1
requests==2.32.3
psutil>=5.9.6
httpx>=0.27.0
